fix(web-ongoing-operations): remove modal for update page
ref: #MANAGER-17981
