+++ packages/manager/apps/web-ongoing-operations/public/translations/dashboard/Messages_fr_FR.json
@@ -100,12 +100,13 @@
   "domain_operations_upload_file_label": "Rechercher les fichiers",
   "domain_operations_upload_dropzone_label": "Glisser & déposer un fichier",
   "domain_operations_upload_accepted_file_types": "Les fichiers aux formats suivants sont acceptés : {{t0}}",
-  "domain_operations_upload_success": "L'upload a été un succès",
+  "domain_operations_upload_success": "L'upload a été un succès, vous allez être redirigé vers le tableau de bord",
   "domain_operations_upload_error": "Echec de l'upload du document",
   "domain_operations_upload_error_with_value": "Echec de l'upload du document : {{t0}}",
   "domain_operations_upload_doing": "Upload en cours...",
   "domain_operations_upload_error_max-file-reached": "Nombre de fichiers maximum atteint",
   "domain_operations_upload_error_size-too-large": "Fichier trop volumineux",
+  "domain_operations_upload_error_wrong-format": "Format de fichier non autorisé",
   "domain_operations_update_cancel": "Opération annulée!",
   "domain_operations_update_relaunch": "Opération relancée!",
   "domain_operations_update_accelerate": "Demande d'accélération de l'opération réussie!",
@@ -117,7 +118,7 @@
   "domain_operations_update_key_identificationNumber": "Mettre à jour le numéro d'identification avec un contenu valide",
   "domain_operations_update_key_owner_email": "Mettre à jour l'email du titulaire avec un contenu valide",
   "domain_operations_update_key_legitimacyAfnic": "Justifier de l'usage du nom de domaine",
-  "domain_operations_modal_title": "Données de l'opération sur {{t0}}",
+  "domain_operations_update_title": "Mise à jour de l'opération pour le domaine {{t0}}",
   "domain_operations_update_me_fields": "Modification du {{t0}}",
   "domain_operations_update_contact_not_me": "Seul(s) le(s) {{t0}} peuvent changer {{t1}}",
   "domain_operations_update_nicadmin_click": "Changez les informations du contact administrateur en cliquant ici",
