+++ packages/manager/apps/web-ongoing-operations/src/components/Update/Content/Update.Me.Contact.component.tsx
@@ -8,7 +8,7 @@ import { useNichandle } from '@/hooks/nichandle/useNichandle';
 import { useGetDomainInformation } from '@/hooks/data/query';
 import Loading from '@/components/Loading/Loading';
 
-interface MeContactComponentProps {
+interface ActionMeContactComponentProps {
   readonly argumentKey: string;
   readonly value: string;
   readonly domainName: string;
@@ -16,13 +16,13 @@ interface MeContactComponentProps {
   readonly fields: string[];
 }
 
-export default function MeContactComponent({
+export default function ActionMeContactComponent({
   argumentKey,
   value,
   domainName,
   operationName,
   fields,
-}: MeContactComponentProps) {
+}: ActionMeContactComponentProps) {
   const { t } = useTranslation('dashboard');
   const { data: webUrl } = useNavigationGetUrl(['web', '', {}]);
   const { data: dedicatedUrl } = useNavigationGetUrl(['dedicated', '', {}]);
@@ -33,20 +33,17 @@ export default function MeContactComponent({
     return <Loading />;
   }
 
-  let url = `${webUrl}/domain/${domainName}/contact-management/edit-contact/${value}/`;
-  if (
-    !serviceInfo &&
-    [domainIncomingTransfer, domainCreate].includes(operationName)
-  ) {
-    url = `${dedicatedUrl}/contact/${value}/${
-      fields.length ? getNicParams(fields) : ''
-    }`;
-  }
-
-  if (nichandle !== serviceInfo?.contactAdmin.id) {
+  if (serviceInfo && nichandle !== serviceInfo.contactAdmin.id) {
     return (
       <OdsText>{t('domain_operations_update_contact_administrator')}</OdsText>
     );
+  }
+
+  let url = `${webUrl}/domain/${domainName}/contact-management/edit-contact/${value}/`;
+  if ([domainIncomingTransfer, domainCreate].includes(operationName)) {
+    url = `${dedicatedUrl}/contact/${value}/${
+      fields.length ? getNicParams(fields) : ''
+    }`;
   }
 
   return (
@@ -60,7 +57,7 @@ export default function MeContactComponent({
       )}
       className="block"
       icon="external-link"
-      data-testid="contactModal"
+      data-testid="contactupdate"
       isDisabled={!url}
     />
   );
