+++ packages/manager/apps/web-ongoing-operations/src/enum/actionName.enum.ts
@@ -1,4 +1,4 @@
-export enum OperationName {
+export enum ActionNameEnum {
   CanCancel = 'cancel',
   CanRelaunch = 'relaunch',
   CanAccelerate = 'accelerate',
