+++ packages/manager/apps/web-ongoing-operations/src/hooks/update/useOperationArguments.tsx
@@ -17,11 +17,7 @@ const getMeTaskArguments = async (
 
   const data = (await Promise.all(promiseArray)) || [];
 
-  const index = data.findIndex((arg) =>
-    ['/me/contact', 'string', '/me'].includes(arg.type),
-  );
-  const actions = data.length === 0 || index >= 0;
-  return { data, actions };
+  return { data };
 };
 
 export function useOperationArguments(id: number) {
