+++ packages/manager/apps/web-ongoing-operations/src/hooks/useOngoingOperationDatagridColumns.tsx
@@ -14,7 +14,6 @@ import { DNS_OPERATIONS_TABLE_HEADER_DOMAIN } from '@/pages/dashboard/Dashboard'
 export const useOngoingOperationDatagridColumns = (
   parent: ParentEnum.DOMAIN | ParentEnum.ZONE,
   data: TOngoingOperations[],
-  openModal: (id: number) => void,
 ) => {
   const { t } = useTranslation('dashboard');
   const l = getI18n();
@@ -93,8 +92,8 @@ export const useOngoingOperationDatagridColumns = (
                 className: `${!props.canAccelerate &&
                   !props.canRelaunch &&
                   !props.canCancel &&
-                  'hidden'} openModal menu-item-button`,
-                onClick: () => openModal(props.id),
+                  'hidden'} menu-item-button`,
+                onClick: () => navigate(`/update/${props.id}`),
               },
               {
                 id: 2,
