+++ packages/manager/apps/web-ongoing-operations/src/index.scss
@@ -15,16 +15,8 @@ html {
   color: var(--ods-color-primary-800);
 }
 
-.modal::part(dialog) {
-  max-height: 70vh;
-}
-
 td > ods-text > ods-link {
   line-height: 1.2rem;
-}
-
-.modal-link::part(label) {
-  white-space: wrap;
 }
 
 .menu-item-button {
