+++ packages/manager/apps/web-ongoing-operations/src/pages/dashboard/domain/Domain.spec.tsx
@@ -1,18 +1,13 @@
 import '@/setupTests';
 import React, { PropsWithChildren } from 'react';
 import '@testing-library/jest-dom';
-import {
-  QueryClient,
-  QueryClientProvider,
-  useQuery,
-} from '@tanstack/react-query';
+import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
 import { vi } from 'vitest';
 import { useResourcesIcebergV6 } from '@ovh-ux/manager-react-components';
-import { render, waitFor, screen, fireEvent } from '@testing-library/react';
+import { render, waitFor, screen } from '@testing-library/react';
 import { domain } from '@/__mocks__/domain';
 import Domain from '@/pages/dashboard/domain/Domain';
 import { taskMeDomain } from '@/constants';
-import { modalOpen } from '@/__mocks__/modal';
 
 vi.mock('react-router-dom', () => ({
   useNavigate: () => vi.fn(() => null),
@@ -133,34 +128,6 @@ describe('Domain datagrid', () => {
 
       const ableButton = buttons[1]; // The button comes from the mock, it's the mock's second element
       expect(ableButton).toHaveAttribute('is-disabled', 'false');
-    });
-  });
-});
-
-describe('Modal Operations domain', () => {
-  it('Display the modal', async () => {
-    (useResourcesIcebergV6 as jest.Mock).mockReturnValue({
-      flattenData: modalOpen,
-      isLoading: false,
-      totalCount: modalOpen.length,
-    });
-
-    (useQuery as jest.Mock).mockReturnValue({
-      data: {
-        data: [],
-        actions: false,
-      },
-    });
-
-    const { container } = render(<Domain />, { wrapper });
-    fireEvent.click(screen.getByTestId('navigation-action-trigger-action'));
-    const openModalButton = container.querySelector('.openModal');
-    fireEvent.click(openModalButton);
-
-    waitFor(() => {
-      expect(
-        screen.getByText('domain_operations_modal_title'),
-      ).toBeInTheDocument();
     });
   });
 });
