+++ packages/manager/apps/web-ongoing-operations/src/pages/dashboard/domain/Domain.tsx
@@ -9,18 +9,10 @@ import Loading from '@/components/Loading/Loading';
 import { useOngoingOperationDatagridColumns } from '@/hooks/useOngoingOperationDatagridColumns';
 import { taskMeDomain } from '@/constants';
 import { TOngoingOperations } from '@/types';
-import Modal from '@/components/Modal/Modal';
-import Notification from '@/components/Notification/Notification.component';
 import { ParentEnum } from '@/enum/parent.enum';
 
 export default function Domain() {
   const { t: tError } = useTranslation('web-ongoing-operations/error');
-  const [modalOpen, setModalOpen] = useState<boolean>(false);
-  const defaultStatus = { label: '', message: '' };
-  const [status, setStatus] = useState(defaultStatus);
-  const [filterDomain, setFilterDomain] = useState<TOngoingOperations | null>(
-    null,
-  );
 
   const {
     flattenData: domainList,
@@ -32,34 +24,15 @@ export default function Domain() {
     sorting,
     setSorting,
     filters,
-    refetch,
   } = useResourcesIcebergV6<TOngoingOperations>({
     route: taskMeDomain,
     queryKey: [taskMeDomain],
     pageSize: 30,
   });
 
-  const openModal = (id: number) => {
-    setModalOpen(true);
-    setFilterDomain(
-      domainList.find((element: TOngoingOperations) => element.id === id),
-    );
-  };
-
-  const closeModal = () => {
-    setModalOpen(!modalOpen);
-    setFilterDomain(null);
-    refetch();
-  };
-
-  const changeStatus = (label: string, message: string) => {
-    setStatus({ label, message });
-  };
-
   const columns = useOngoingOperationDatagridColumns(
     ParentEnum.DOMAIN,
     domainList,
-    openModal,
   );
 
   if (isLoading) {
@@ -83,23 +56,6 @@ export default function Domain() {
 
   return (
     <React.Suspense>
-      {modalOpen && (
-        <Modal
-          universe="domain"
-          onCloseModal={closeModal}
-          operation={filterDomain}
-          changeStatus={changeStatus}
-        />
-      )}
-
-      {status.label && (
-        <Notification
-          label={status.label}
-          message={status.message}
-          removeMessage={() => setStatus(defaultStatus)}
-        />
-      )}
-
       {domainList && (
         <div data-testid="datagrid">
           <Datagrid
