+++ packages/manager/apps/web-ongoing-operations/src/routes/routes.constant.ts
@@ -3,6 +3,6 @@ export const urls = {
   domain: 'domain',
   dns: 'dns',
   track: 'tracking/:id',
-  upload: 'upload/:id',
+  update: 'update/:id',
   error404: '404',
 };
